backend/requirements.txt
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-dotenv==0.21.1
